Refactor packages
Not supposed to edit shoppingLists package under store
