+++ src/store/JasonYbReduxShoppingList/actions.ts
@@ -1,0 +1,18 @@
+import { Dispatch, GetState } from '../types'
+import { ActionTypes, SHOPPING_LIST_ACTIONS, ShoppingItem } from './types'
+
+//set my 'exampleList' state with my argument 'newList'
+//to call this in your component or page, wrap with dispatch
+//eg dispatch(mockActionSetMyExampleList(theNewList))
+
+export const AddItemQuantity =
+  (itemToChange: ShoppingItem) => async (dispatch: Dispatch<ActionTypes>, getState: GetState) => {
+    const { itemList } = getState().JasonYbReduxShoppingList
+    const updatedList = itemList.map((item) =>
+      item.index === itemToChange.index ? { ...item, quantity: item.quantity + 1 } : item,
+    )
+    dispatch({
+      type: SHOPPING_LIST_ACTIONS.SET_ITEM_LIST,
+      itemList: updatedList,
+    })
+  }
